Add raw middleware for auth
