+++ auth-service/.eslintrc.yml
