+++ auth-service/Dockerfile
@@ -7,5 +7,5 @@ RUN npm install
 
 COPY . .
 
-EXPOSE 8000
+EXPOSE 7000
 CMD [ "npm", "start" ]
