+++ auth-service/cache/setup.js
