+++ auth-service/constants/httpResponse.js
