+++ auth-service/package.json
